moved VerticalGapData to separate class
