+++ src/engraving/libmscore/layout.cpp
@@ -1970,229 +1970,4 @@ void Score::doLayoutRange(const Fraction& st, const Fraction& et)
 {
     m_layout.doLayoutRange(st, et);
 }
-
-//---------------------------------------------------------
-//   VerticalStretchData
-//      defines a gap ABOVE the staff.
-//---------------------------------------------------------
-
-VerticalGapData::VerticalGapData(bool first, System* sys, Staff* st, SysStaff* sst, Spacer* nextSpacer, qreal y)
-    : _fixedHeight(first), system(sys), sysStaff(sst), staff(st)
-{
-    if (_fixedHeight) {
-        _normalisedSpacing = system->score()->styleP(Sid::staffUpperBorder);
-        _maxActualSpacing = _normalisedSpacing;
-    } else {
-        _normalisedSpacing = system->y() + (sysStaff ? sysStaff->bbox().y() : 0.0) - y;
-        _maxActualSpacing = system->score()->styleP(Sid::maxStaffSpread);
-
-        Spacer* spacer { staff ? system->upSpacer(staff->idx(), nextSpacer) : nullptr };
-
-        if (spacer) {
-            _fixedSpacer = spacer->spacerType() == SpacerType::FIXED;
-            _normalisedSpacing = qMax(_normalisedSpacing, spacer->gap());
-            if (_fixedSpacer) {
-                _maxActualSpacing = _normalisedSpacing;
-            }
-        }
-    }
-}
-
-//---------------------------------------------------------
-//   updateFactor
-//---------------------------------------------------------
-
-void VerticalGapData::updateFactor(qreal factor)
-{
-    if (_fixedHeight) {
-        return;
-    }
-    qreal f = qMax(factor, _factor);
-    _normalisedSpacing *= _factor / f;
-    _factor = f;
-}
-
-//---------------------------------------------------------
-//   addSpaceBetweenSections
-//---------------------------------------------------------
-
-void VerticalGapData::addSpaceBetweenSections()
-{
-    updateFactor(system->score()->styleD(Sid::spreadSystem));
-    if (!(_fixedHeight | _fixedSpacer)) {
-        _maxActualSpacing = system->score()->styleP(Sid::maxSystemSpread) / _factor;
-    }
-}
-
-//---------------------------------------------------------
-//   addSpaceAroundVBox
-//---------------------------------------------------------
-
-void VerticalGapData::addSpaceAroundVBox(bool above)
-{
-    _fixedHeight = true;
-    _factor = 1.0;
-    const Score* score { system->score() };
-    _normalisedSpacing = above ? score->styleP(Sid::frameSystemDistance) : score->styleP(Sid::systemFrameDistance);
-    _maxActualSpacing = _normalisedSpacing / _factor;
-}
-
-//---------------------------------------------------------
-//   addSpaceAroundNormalBracket
-//---------------------------------------------------------
-
-void VerticalGapData::addSpaceAroundNormalBracket()
-{
-    updateFactor(system->score()->styleD(Sid::spreadSquareBracket));
-}
-
-//---------------------------------------------------------
-//   addSpaceAroundCurlyBracket
-//---------------------------------------------------------
-
-void VerticalGapData::addSpaceAroundCurlyBracket()
-{
-    updateFactor(system->score()->styleD(Sid::spreadCurlyBracket));
-}
-
-//---------------------------------------------------------
-//   insideCurlyBracket
-//---------------------------------------------------------
-
-void VerticalGapData::insideCurlyBracket()
-{
-    _maxActualSpacing = system->score()->styleP(Sid::maxAkkoladeDistance) / _factor;
-}
-
-//---------------------------------------------------------
-//   factor
-//---------------------------------------------------------
-
-qreal VerticalGapData::factor() const
-{
-    return _factor;
-}
-
-//---------------------------------------------------------
-//   spacing
-//    return normalised spacing
-//---------------------------------------------------------
-
-qreal VerticalGapData::spacing() const
-{
-    return _normalisedSpacing + _addedNormalisedSpace;
-}
-
-//---------------------------------------------------------
-//   addedSpace
-//---------------------------------------------------------
-
-qreal VerticalGapData::actualAddedSpace() const
-{
-    return _addedNormalisedSpace * factor();
-}
-
-//---------------------------------------------------------
-//   addSpacing
-//---------------------------------------------------------
-
-qreal VerticalGapData::addSpacing(qreal step)
-{
-    if (_fixedHeight | _fixedSpacer) {
-        return 0.0;
-    }
-    if (_normalisedSpacing >= _maxActualSpacing) {
-        _normalisedSpacing = _maxActualSpacing;
-        step = 0.0;
-    } else {
-        qreal newSpacing { _normalisedSpacing + _addedNormalisedSpace + step };
-        if ((newSpacing >= _maxActualSpacing)) {
-            step = _maxActualSpacing - _normalisedSpacing - _addedNormalisedSpace;
-        }
-    }
-    _addedNormalisedSpace += step;
-    _lastStep = step;
-    return step;
-}
-
-//---------------------------------------------------------
-//   isFixedHeight
-//---------------------------------------------------------
-
-bool VerticalGapData::isFixedHeight() const
-{
-    return _fixedHeight || almostZero(_normalisedSpacing - _maxActualSpacing);
-}
-
-//---------------------------------------------------------
-//   undoLastAddSpacing
-//---------------------------------------------------------
-
-void VerticalGapData::undoLastAddSpacing()
-{
-    _addedNormalisedSpace -= _lastStep;
-    _lastStep = 0.0;
-}
-
-//---------------------------------------------------------
-//   addFillSpacing
-//---------------------------------------------------------
-
-qreal VerticalGapData::addFillSpacing(qreal step, qreal maxFill)
-{
-    if (_fixedSpacer) {
-        return 0.0;
-    }
-    qreal actStep { ((step + _fillSpacing / _factor) > maxFill) ? (maxFill - _fillSpacing / _factor) : step };
-    qreal res = addSpacing(actStep);
-    _fillSpacing += res * _factor;
-    return res;
-}
-
-//---------------------------------------------------------
-//   deleteAll
-//---------------------------------------------------------
-
-void VerticalGapDataList::deleteAll()
-{
-    for (auto vsd : *this) {
-        delete vsd;
-    }
-}
-
-//---------------------------------------------------------
-//   sumStretchFactor
-//---------------------------------------------------------
-
-qreal VerticalGapDataList::sumStretchFactor() const
-{
-    qreal sum { 0.0 };
-    for (VerticalGapData* vsd : *this) {
-        if (!vsd->isFixedHeight()) {
-            sum += vsd->factor();
-        }
-    }
-    return sum;
-}
-
-//---------------------------------------------------------
-//   smallest
-//---------------------------------------------------------
-
-qreal VerticalGapDataList::smallest(qreal limit) const
-{
-    VerticalGapData* vdp { nullptr };
-    for (VerticalGapData* vgd : *this) {
-        if (vgd->isFixedHeight()) {
-            continue;
-        }
-        if ((qCeil(limit) == qCeil(vgd->spacing()))) {
-            continue;
-        }
-        if (!vdp || (vgd->spacing() < vdp->spacing())) {
-            vdp = vgd;
-        }
-    }
-    return vdp ? vdp->spacing() : 0.0;
-}
 }
