+++ src/engraving/libmscore/libmscore.cmake
@@ -357,4 +357,6 @@ set(LIBMSCORE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/layout/layoutmeasure.h
     ${CMAKE_CURRENT_LIST_DIR}/layout/layoutbeams.cpp
     ${CMAKE_CURRENT_LIST_DIR}/layout/layoutbeams.h
+    ${CMAKE_CURRENT_LIST_DIR}/layout/verticalgapdata.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/layout/verticalgapdata.h
 )
